requirements.txt
django>=5.0
djangorestframework==3.15
pyodbc>= 5.1
django-filter>=24.2
drf-spectacular>=0.27
psycopg2-binary>=2.9
django-cors-headers>=4.3
requests>=2.32